backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0.post1
motor==3.3.2
redis==5.0.1
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
python-dotenv==1.0.0
structlog==24.1.0
boto3==1.34.34
azure-mgmt-resource==23.1.0
azure-identity==1.15.0
